feat: abstract storage
