+++ app/db.py
@@ -1,15 +1,8 @@
 from logging import getLogger
-from pathlib import Path
 from typing import List, Mapping, Optional, Type
+from abc import ABC, abstractmethod
 
-from html2text import HTML2Text
-from markdown2 import markdown
-from readabilipy import simple_json_from_html_string
-from readability import Document
-from tinydb import Query, TinyDB
-
-from app.constants import DATA_DIR
-from app.content.engine import ContentRetrievalEngine, content_handler
+from app.content.engine import ContentRetrievalEngine
 from app.models import (
     ContentRetrievalHandler,
     EntryContent,
@@ -20,18 +13,13 @@ from app.models import (
     SummarizationHandler,
 )
 from app.notification.engine import NotificationEngine
-from app.summarization.engine import SummarizationEngine, summarization_handler
+from app.summarization.engine import SummarizationEngine
+from app.reading import ReadingMethodsMixIn
 
-logger = getLogger("uvicorn.error")
 
+class StorageHandler(ABC, ReadingMethodsMixIn):
 
-class DB:
-    """
-    Use this class to encapsulate DB interactions
-    """
-
-    db_path = Path(DATA_DIR, "db.json").resolve()
-    db = TinyDB(db_path)
+    logger = getLogger("uvicorn.error")
 
     handler_map = {
         **SummarizationEngine.handlers,
@@ -51,221 +39,93 @@ class DB:
         **{k: "content" for k in ContentRetrievalEngine.handlers.keys()},
     }
 
+    def reconfigure_handler(self, id: str, config: Mapping):
+
+        return self.engine_map[self.handler_type_map[id]](
+            type=id, config=config
+        ).get_handler()
+    
+    @abstractmethod
     def clear_active_feeds(self) -> None:
-        self.db.drop_table("feeds")
+        pass
+    
+    @abstractmethod
+    def insert_feed(self, feed: Feed) -> None:
+        pass
 
-    def insert_feed(self, feed: Feed):
-        table = self.db.table("feeds")
-        table.insert({"id": feed.id, "feed": feed.dict()})
-
+    @abstractmethod
     def get_feed(self, id: str) -> Feed:
-        table = self.db.table("feeds")
-
-        query = Query()
-        feed = table.search(query.id == id)[0]["feed"]
-
-        return Feed(**feed)
-
+        pass
+    
+    @abstractmethod
     def get_feeds(self) -> List[Feed]:
-        table = self.db.table("feeds")
-
-        return [Feed(**i["feed"]) for i in table.all()]
-
+        pass
+    
+    @abstractmethod
     def get_poll_state(self, feed: Feed) -> Optional[int]:
-        table = self.db.table("poll")
-
-        query = Query().id.matches(feed.id)
-
-        results = table.search(query)
-
-        if results:
-            return results[0]["last_polled_at"]
-
-    def set_feed_start_ts(self, feed: Feed, start_ts: int):
-        table = self.db.table("feed_start")
-
-        query = Query().id.matches(feed.id)
-        table.upsert({"id": feed.id, "start_ts": start_ts}, cond=query)
-
+        pass
+    
+    @abstractmethod
+    def set_feed_start_ts(self, feed: Feed, start_ts: int) -> None:
+        pass
+    
+    @abstractmethod
     def get_feed_start_ts(self, feed: Feed) -> int:
-        table = self.db.table("feed_start")
+        pass
+    
+    @abstractmethod
+    def update_poll_state(self, feed: Feed, now: int) -> None:
+        pass
+    
+    @abstractmethod
+    def upsert_feed_entry(self, feed: Feed, entry: FeedEntry) -> None:
+        pass
+    
+    @abstractmethod
+    def get_entries(self, feed: Feed = None) -> Mapping[str, FeedEntry | str]:
+        pass
+    
+    @abstractmethod
+    def get_feed_entry(self, id: str) -> FeedEntry:
+        pass
+    
+    @abstractmethod
+    def feed_entry_exists(self, id: str) -> bool:
+        pass
 
-        query = Query().id.matches(feed.id)
-        results = table.search(query)
-
-        if results:
-            return results[0]["start_ts"]
-
-    def update_poll_state(self, feed: Feed, now: int):
-        table = self.db.table("poll")
-
-        query = Query().id.matches(feed.id)
-        table.upsert({"id": feed.id, "last_polled_at": now}, cond=query)
-
-    def upsert_feed_entry(self, feed: Feed, entry: FeedEntry):
-        table = self.db.table("entries")
-
-        row = {
-            "id": entry.id,
-            "feed_id": feed.id,
-            "entry": entry.dict(),
-        }
-
-        query = Query().id.matches(entry.id)
-        table.upsert(row, cond=query)
-
-    def get_entries(self, feed: Feed = None):
-        table = self.db.table("entries")
-
-        if feed:
-            query = Query().feed_id.matches(feed.id)
-            entries = table.search(query)
-        else:
-            entries = table.all()
-
-        return [
-            {"entry": FeedEntry(**i["entry"]), "feed_id": i["feed_id"], "id": i["id"]}
-            for i in entries
-        ]
-
-    def get_feed_entry(self, id: str):
-        table = self.db.table("entries")
-
-        query = Query().id.matches(id)
-        entry = table.search(query)[0]
-
-        return FeedEntry(**entry["entry"])
-
-    def feed_entry_exists(self, id: str):
-        table = self.db.table("entries")
-
-        query = Query().id.matches(id)
-        if table.search(query):
-            return True
-        else:
-            return False
-
-    @staticmethod
-    async def _get_entry_html(url: str) -> str:
-
-        return await content_handler.get_content(url)
-
-    @staticmethod
-    def _get_entry_md(content: str) -> str:
-
-        md = simple_json_from_html_string(html=content)
-
-        cleaned_document = Document(input=md["content"])
-
-        converter = HTML2Text()
-        converter.ignore_images = True
-        converter.ignore_links = True
-
-        return converter.handle(cleaned_document.summary(html_partial=True))
-
+    @abstractmethod
     async def get_entry_content(
         self, entry: FeedEntry, redrive: bool = False
     ) -> EntryContent:
-        table = self.db.table("entry_contents")
-        query = Query().id.matches(entry.id)
+        pass
 
-        existing = table.search(query)
-        if existing and not redrive:
-            return EntryContent(**existing[0]["entry_contents"])
-
-        else:
-            if redrive:
-                logger.info(f"starting redrive for feed entry {entry.id}")
-
-            raw_content = await self._get_entry_html(entry.url)
-            content = self._get_entry_md(content=raw_content)
-
-            summary = summarization_handler.summarize(
-                feed=self.get_feed(entry.feed_id), entry=FeedEntry, mk=content
-            )
-
-            entry_content = EntryContent(
-                url=entry.url,
-                content=markdown(content),
-                summary=markdown(summary) if summary else None,
-            )
-
-            query = Query().id.matches(entry.id)
-            table.upsert(
-                {"id": entry_content.id, "entry_contents": entry_content.dict()},
-                cond=query,
-            )
-
-            return entry_content
-
+    @abstractmethod
     def upsert_handler(
         self,
         handler: Type[
             SummarizationHandler | NotificationHandler | ContentRetrievalHandler
         ],
     ) -> None:
-        table = self.db.table("handler")
-
-        row = {
-            "id": handler.id,
-            "handler": handler.dict(),
-        }
-
-        query = Query().id.matches(handler.id)
-        table.upsert(row, cond=query)
-
-    def _make_handler_obj(self, id: str, config: Mapping):
-
-        return self.engine_map[self.handler_type_map[id]](
-            type=id, config=config
-        ).get_handler()
-
+        pass
+    
+    @abstractmethod
     def get_handlers(
         self,
     ) -> Mapping[
         str, Type[SummarizationHandler | NotificationHandler | ContentRetrievalHandler]
     ]:
-        table = self.db.table("handler")
-
-        handlers = {i: None for i in self.handler_map.keys()}
-
-        for cfg in table.all():
-            handlers[cfg["id"]] = self._make_handler_obj(
-                id=cfg["id"], config=cfg["handler"]
-            )
-
-        return handlers
-
+        pass
+    
+    @abstractmethod
     def get_handler(
         self, id: str
     ) -> Type[SummarizationHandler | NotificationHandler | ContentRetrievalHandler]:
-        table = self.db.table("handler")
+        pass
 
-        query = Query().id.matches(id)
-        handler = table.search(query)[0]
-
-        handler_obj = self._make_handler_obj(
-            id=handler["id"], config=handler["handler"]
-        )
-
-        return handler_obj
-
+    @abstractmethod
     def get_settings(self) -> GlobalSettings:
-        table = self.db.table("settings")
-
-        try:
-            settings = table.all()[0]
-            return GlobalSettings(**settings["settings"])
-        except IndexError:
-            return GlobalSettings()
-
+        pass
+    
+    @abstractmethod
     def upsert_settings(self, settings: GlobalSettings) -> None:
-        table = self.db.table("settings")
-
-        row = {
-            "id": "settings",
-            "settings": settings.dict(),
-        }
-
-        query = Query().id.matches("settings")
-        table.upsert(row, cond=query)
+        pass
