+++ app/handlers.py
@@ -1,9 +1,7 @@
-from app.db import DB
+from app.storage.engine import storage_handler as db
 from app.summarization.engine import summarization_handler
 from app.notification.engine import notification_handler
 from app.content.engine import content_handler
-
-db = DB()
 
 def load_handlers():
     db.upsert_handler(summarization_handler)
